+++ src/index.js
@@ -2,10 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
 import * as serviceWorker from './serviceWorker';
-import TSChart from './TSChart';
-import HistogramChart from './HistogramChart';
-import News from './News';
-import Stats from './Stats';
+
 
 import stockApp from './reducers/reducers'
 import App from './App'
@@ -27,10 +24,6 @@ let createStoreWithMiddleware = applyMiddleware(thunkMiddleware)(createStore)
 
 let store = createStoreWithMiddleware(stockApp)
 
-//ReactDOM.render(<TSChart />, document.getElementById('root'));
-ReactDOM.render(<HistogramChart />, document.getElementById('hist'));
-ReactDOM.render(<Stats />, document.getElementById('stats'));
-ReactDOM.render(<News />, document.getElementById('news'));
 
 ReactDOM.render(
     <Provider store={store}>
