+++ src/reducers/profileReducer.js
@@ -10,28 +10,23 @@ import {
       isAuthenticated: localStorage.getItem('id_token') ? true : false
     }, action) {
     switch (action.type) {
-      case LOGIN_REQUEST:
+      case PROFILE_GET:
         return Object.assign({}, state, {
           isFetching: true,
           isAuthenticated: false,
           user: action.creds
         })
-      case LOGIN_SUCCESS:
+      case PROFILE_GET_SUCCESS:
         return Object.assign({}, state, {
           isFetching: false,
           isAuthenticated: true,
           errorMessage: ''
         })
-      case LOGIN_FAILURE:
+      case PROFILE_GET_FAILURE:
         return Object.assign({}, state, {
           isFetching: false,
           isAuthenticated: false,
           errorMessage: action.message
-        })
-      case LOGOUT_SUCCESS:
-        return Object.assign({}, state, {
-          isFetching: true,
-          isAuthenticated: false
         })
       default:
         return state
